Renamed MenuItem to MenuAction
