+++ src/main/com/app/App.java
@@ -1,13 +1,13 @@
 package com.app;
 
 import com.app.controller.Menu;
-import com.app.model.menuItems.*;
+import com.app.model.menuActions.*;
 
 public class App {
   public static void main(String[] args) {
-    createOrderMenuItem test2 = new createOrderMenuItem();
-    exitMenuItem test3 = new exitMenuItem();
-    MenuItem[] menu = {test2,test3};
+    createOrderMenuAction test2 = new createOrderMenuAction();
+    exitMenuAction test3 = new exitMenuAction();
+    MenuAction[] menu = {test2,test3};
     Menu test = new Menu("Hello", "Hello", menu);
 
     test.run();
