+++ src/main/com/app/controller/Menu.java
@@ -1,7 +1,7 @@
 package com.app.controller;
 
 import com.app.model.MenuModel;
-import com.app.model.menuItems.MenuItem;
+import com.app.model.menuActions.MenuAction;
 import com.app.view.MenuView;
 
 import java.util.Scanner;
@@ -11,8 +11,8 @@ public class Menu {
   private Scanner scanner = new Scanner(System.in);
   private MenuView menuView;
 
-  public Menu(String menuHeader, String leadText, MenuItem[] menuItems) {
-    menu = new MenuModel(menuHeader, menuItems, leadText);
+  public Menu(String menuHeader, String leadText, MenuAction[] menuActions) {
+    menu = new MenuModel(menuHeader, menuActions, leadText);
     menuView = new MenuView();
   }
