+++ src/main/com/app/model/MenuModel.java
@@ -1,20 +1,20 @@
 package com.app.model;
 
-import com.app.model.menuItems.*;
+import com.app.model.menuActions.*;
 
 public class MenuModel {
-  private MenuItem[] menuItems;
+  private MenuAction[] menuActions;
   private String menuHeader;
   private String leadText;
 
-  public MenuModel(String menuHeader, MenuItem[] menuItems, String leadText){
-    this.menuItems = menuItems;
+  public MenuModel(String menuHeader, MenuAction[] menuActions, String leadText){
+    this.menuActions = menuActions;
     this.menuHeader = menuHeader;
     this.leadText = leadText;
   }
 
-  public MenuItem getMenuItem(int index){
-    return menuItems[index];
+  public MenuAction getMenuItem(int index){
+    return menuActions[index];
   }
   public String getLeadText(){
     return this.leadText;
