+++ src/main/com/app/model/menuActions/createOrderMenuAction.java
@@ -1,0 +1,10 @@
+package com.app.model.menuActions;
+
+public class createOrderMenuAction extends MenuAction {
+
+  @Override
+  public void run() {
+    System.out.println("Test");
+  }
+
+}
