+++ src/main/com/app/model/menuActions/exitMenuAction.java
@@ -1,0 +1,13 @@
+package com.app.model.menuActions;
+
+public class exitMenuAction extends MenuAction {
+
+  public exitMenuAction(){
+    super(false);
+  }
+
+  @Override
+  public void run() {
+    System.out.println("Exit");
+  }
+}
