--- src/main/com/app/model/menuItems/createOrderMenuItem.java
@@ -1,10 +1,0 @@
-package com.app.model.menuItems;
-
-public class createOrderMenuItem extends MenuItem{
-
-  @Override
-  public void run() {
-    System.out.println("Test");
-  }
-
-}
